requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.4

# Task Queue
celery==5.3.4
redis==5.0.1

# Scrapers
praw==7.7.1  # Reddit
snscrape==0.7.0.20230622  # Twitter
apify-client==1.7.1  # Apify platform
beautifulsoup4==4.12.2
requests==2.31.0

# AI/ML
anthropic==0.7.7  # Claude API
openai==1.3.7  # GPT + embeddings
httpx==0.25.2  # For Cloudflare AI

# Agent Framework
agno==0.1.0  # Agno AI agents

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2
tenacity==8.2.3  # Retry logic
